#25 [mod] background translucent 코드 수정
