+++ app/src/main/java/com/sopetit/softie/ui/onboarding/themechoice/ChoiceThemeFragment.kt
@@ -1,16 +1,13 @@
 package com.sopetit.softie.ui.onboarding.themechoice
 
 import android.os.Bundle
-import android.text.SpannableStringBuilder
-import android.text.Spanned
-import android.text.style.ForegroundColorSpan
 import android.view.View
-import androidx.core.content.ContextCompat
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.GridLayoutManager
 import com.sopetit.softie.R
 import com.sopetit.softie.databinding.FragmentOnboardingChoiceThemeBinding
+import com.sopetit.softie.ui.onboarding.OnboardingActivity.Companion.MAXIMUM_THEME_SELECTION
 import com.sopetit.softie.ui.onboarding.OnboardingViewModel
 import com.sopetit.softie.util.binding.BindingFragment
 
@@ -31,37 +28,8 @@ class ChoiceThemeFragment :
         binding.viewModel = viewModel
         binding.themeViewModel = themeViewModel
 
-        initSetTranslucentBackground()
-        initSetSpeechText()
         initMakeThemeAdapter()
         initChangeFragment()
-    }
-
-    private fun initSetTranslucentBackground() {
-        binding.clOnboardingChoiceThemeTranslucentBackground.setOnClickListener {
-            themeViewModel.setLayoutTranslucent(false)
-            viewModel.setLayoutTranslucent(false)
-        }
-    }
-
-    private fun initSetSpeechText() {
-        val nickname = viewModel.bearNickname.value ?: ""
-        val message = getString(R.string.onboarding_choice_theme_speech).format(nickname)
-
-        binding.tvOnboardingChoiceThemeSpeech.text =
-            SpannableStringBuilder(message).apply {
-                setSpan(
-                    ForegroundColorSpan(
-                        ContextCompat.getColor(
-                            requireActivity(),
-                            R.color.onboarding_speech
-                        )
-                    ),
-                    SPAN_START,
-                    SPAN_START + nickname.length,
-                    Spanned.SPAN_EXCLUSIVE_EXCLUSIVE
-                )
-            }
     }
 
     private fun initMakeThemeAdapter() {
@@ -107,10 +75,5 @@ class ChoiceThemeFragment :
     override fun onDestroyView() {
         super.onDestroyView()
         _choiceThemeAdapter = null
-    }
-
-    companion object {
-        const val SPAN_START = 5
-        const val MAXIMUM_THEME_SELECTION = 3
     }
 }
