+++ app/src/main/res/layout/activity_onboarding.xml
@@ -74,7 +74,7 @@
         </androidx.constraintlayout.widget.ConstraintLayout>
 
         <androidx.constraintlayout.widget.ConstraintLayout
-            android:id="@+id/cl_onboarding_choice_theme_translucent_background"
+            android:id="@+id/cl_onboarding_theme_translucent_background"
             android:layout_width="0dp"
             android:layout_height="0dp"
             android:background="@color/onboarding_translucent"
@@ -92,5 +92,58 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/cl_onboarding_top_bar" />
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:id="@+id/cl_onboarding_theme_translucent_background_content"
+            android:layout_width="0dp"
+            android:layout_height="0dp"
+            android:background="@color/onboarding_translucent"
+            android:visibility="@{viewModel.layoutTranslucent ? View.VISIBLE : View.GONE}"
+            app:layout_constraintTop_toTopOf="@id/fcv_onboarding_fragment"
+            app:layout_constraintStart_toStartOf="@id/fcv_onboarding_fragment"
+            app:layout_constraintEnd_toEndOf="@id/fcv_onboarding_fragment"
+            app:layout_constraintBottom_toBottomOf="@id/fcv_onboarding_fragment">
+
+            <ImageView
+                android:id="@+id/iv_onboarding_theme_title_bear"
+                android:layout_width="53dp"
+                android:layout_height="50dp"
+                android:src="@drawable/ic_bear_face_1"
+                android:layout_marginTop="34dp"
+                android:layout_marginStart="22dp"
+                android:visibility="@{viewModel.layoutTranslucent ? View.VISIBLE : View.GONE}"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintStart_toStartOf="parent" />
+
+            <ImageView
+                android:id="@+id/iv_onboarding_theme_speech_background"
+                android:layout_width="0dp"
+                android:layout_height="60dp"
+                android:layout_marginTop="29dp"
+                android:layout_marginStart="14dp"
+                android:layout_marginEnd="23dp"
+                android:background="@drawable/ic_speech_long"
+                android:visibility="@{viewModel.layoutTranslucent ? View.VISIBLE : View.GONE}"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintStart_toEndOf="@id/iv_onboarding_theme_title_bear"
+                app:layout_constraintEnd_toEndOf="parent" />
+
+            <TextView
+                android:id="@+id/tv_onboarding_theme_speech"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginVertical="14dp"
+                android:text="@{viewModel.layoutTranslucent ? @string/onboarding_choice_theme_speech(viewModel.bearNickname) : @string/onboarding_choice_theme_speech_after}"
+                android:visibility="@{viewModel.layoutTranslucent ? View.VISIBLE : View.GONE}"
+                android:textColor="@color/gray700"
+                android:textAppearance="@style/bubble16"
+                android:bufferType="spannable"
+                app:layout_constraintTop_toTopOf="@id/iv_onboarding_theme_speech_background"
+                app:layout_constraintBottom_toBottomOf="@id/iv_onboarding_theme_speech_background"
+                app:layout_constraintStart_toStartOf="@id/iv_onboarding_theme_speech_background"
+                app:layout_constraintEnd_toEndOf="@id/iv_onboarding_theme_speech_background" />
+
+        </androidx.constraintlayout.widget.ConstraintLayout>
+
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
